Relatorio de veiculos alocados

+++ utils/apuracao_xml.py
@@ -42,7 +42,7 @@ def get_percentual_apuracao(locais, xml):
             dicionario['secoes'] = int(cidade[1]['secoes']) + int(cidade[1]['secoesVT'])
             dicionario['secoes_totalizadas'] = float(abrangencia.attrib['secoesTotalizadas'])
             dicionario['secoes_restantes'] = int(abrangencia.attrib['secoesNaoTotalizadas'])
-            dicionario['percentual'] = float(dicionario['secoes_totalizadas']*100) / dicionario['secoes']
+            dicionario['percentual'] = (dicionario['secoes_totalizadas']*100) / dicionario['secoes']
             dicionario['dt_atualizacao'] = parser.parse('%s %s'%(abrangencia.attrib['dataTotalizacao'], abrangencia.attrib['horaTotalizacao']) )
             dicionario['finalizado'] = abrangencia.attrib['totalizacaoFinal'] == 'S' and True or False
             dicionario['turno'] = int(xml.attrib['turno'])
